Initial Commit (with R Studio's Shiny app setup)
This is a brand new Shiny project from R Studio with its appropriate hidden files. It is our objective to build an application that uses R to run a search engine for images of bones from a database with a set number of parameter describing each image in an attached metadata file.